requirements.txt
streamlit>=1.22.0
pillow>=9.0.0
reportlab>=3.6.12
markdown>=3.4.0
